Fix sentinel master configuration
The master sentinel configuration use "127.0.0.1" to monitor the redis
master (this configuration is active only if the index of the
statefulset is 0). The other sentinel use the DNS to resolve the redis
master IP.

The bug occur when the sentinel run the command fix-config and say to
the redis slave "the new master is 127.0.0.1", after the that the
replication connection is lost by all the redis slave.

To fix this issue, now the master sentinel configuration use the POD_ID
instead of "127.0.0.1" to join the master redis.